Move backend search code into separate files
