+++ civictechprojects/urls.py
@@ -67,11 +67,11 @@ urlpatterns = [
         r'^delete_s3/(?P<s3_key>.*)$',
         views.delete_uploaded_file,
     ),
-    url(r'^api/projects/recent', views.recent_projects_list),
-    url(r'^api/projects', views.projects_list),
+    url(r'^api/projects/recent', views.recent_projects),
+    url(r'^api/projects', views.project_search),
     url(r'^api/events', views.events_list),
     url(r'^api/limited_listings', views.limited_listings),
-    url(r'^api/groups', views.groups_list),
+    url(r'^api/groups', views.group_search),
     url(r'^api/tags/groups', views.group_tags_counts),
     url(r'^api/tags', views.tags),
     url(r'^admin/', admin.site.urls),
